requirements-dev.txt
aws-lambda-typing==1.0.3
boto3-stubs[s3,sqs]==1.17.73
tox==3.23.1
